==> 1 - Read and Save Colors.ipynb <==
import csv

# read colors
mydict = {}
filename = "colors.csv"

with open(filename, mode='r') as inp:
    reader = csv.reader(inp)
    dict_from_csv = {rows[0]:rows[1] for rows in reader}

print(dict_from_csv)

# this is for saving dictionary
        
with open('data.csv', 'w') as f:  
    writer = csv.writer(f)
    for k, v in dict_from_csv.items():
       writer.writerow([k, v])
